requirements.txt
Django
psycopg2
docker
docker-compose
django-ipware

